requirements.txt
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
snowflake==0.0.4
tzdata==2024.1